Add day 13 part 2 solution.
I was able to implement mergesort pretty well.  I'm not sure that
quicksort can be implemented tacitly, since it would require
modifying the array in place.  It's possible.  Though I think I'd
have to be able to inspect the memory somehow in order to know for
sure that it would work.