Day 15 solution.
Doesn't currently work, because it runs out of memory for the actual
input.  It turns out, storing a string grid of that size is far too
large to actually fit in the default memory for J.

I'll have to see if there's a way to increase the memory.  If not, I'll
have to try prunning the input to make sure I don't render unecessary
things (that would let me prune some of the sensors & beacons, then.)

If I can't do that, then I'll need a different solution.